ai/chatbot/requirements.txt
faiss-cpu
sentence-transformers
flask
ollama
numpy
